+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Bugfixes
 
  - Fix missing netcat apt install candidate with package netcat-openbsd
+ - Fix pip requiring --break-system-packages to install/update non-system packages
 
 ### Miscellaneous